+++ cita-chain/types/src/errors/mod.rs
@@ -1,0 +1,3 @@
+mod error;
+
+pub use error::*;
